requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
mysql-connector-python==9.4.0
pydantic==2.11.9
python-dotenv==1.0.1
